requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
oauthlib==2.0.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22
python-dateutil
ipdb==0.9.0
